# public/openapi.yaml
openapi: 3.1.0
info:
  title: Nectar Engine API
  version: 1.0.0
  description: >
    Affiliate workflow API for offer-to-content generation and platform-aware
    compliance scanning. Exposes two operations: generate a campaign toolkit
    from raw affiliate-offer text, and scan promotional copy for platform-
    specific risk signals.
  contact:
    name: Nectar Engine
    url: https://nectar-engine.vercel.app/contact
    email: sinaloainspireddreams@gmail.com
servers:
  - url: https://nectar-engine.vercel.app
    description: Production
paths:
  /api/generate:
    summary: Generate a campaign toolkit from an affiliate offer.
    options:
      summary: CORS preflight
      description: Returns CORS headers for browser-based callers (e.g. ChatGPT Custom GPT Actions).
      operationId: generatePreflight
      responses:
        '204':
          description: No content with CORS headers
    post:
      operationId: generateNectarToolkit
      summary: Turn an affiliate offer into campaign content
      description: >
        Generate campaign angles, nine platform-specific posts, headlines,
        body copy, CTAs, and source-grounded compliance notes. Optionally
        fetches verified offer-page context from a supplied URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
            examples:
              minimal:
                value:
                  offerText: "Offer Name: Acme Widget\nVertical: Software\nPayout: $30 CPA"
              withUrl:
                value:
                  offerText: "Offer Name: Acme Widget\nVertical: Software\nPayout: $30 CPA"
                  offerUrl: "https://example.com/acme-widget"
      responses:
        '200':
          description: Generated toolkit
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/GeneratedToolkit'
        '400':
          description: Invalid offer input — missing or below 20-char minimum
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Upstream provider rate limit or quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Model returned invalid JSON or missing required fields/platforms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Generator not configured (OPENROUTER_API_KEY missing) or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/scan:
    summary: Scan promotional copy for platform-specific risk signals.
    options:
      summary: CORS preflight
      description: Returns CORS headers for browser-based callers (e.g. ChatGPT Custom GPT Actions).
      operationId: scanPreflight
      responses:
        '204':
          description: No content with CORS headers
    post:
      operationId: scanNectarContent
      summary: Scan content for platform-specific risk signals
      description: >
        Return Pass, Warning, or Fail for each selected platform, with exact
        flagged phrases (verbatim substrings of the input) and a safer rewrite
        when needed. Flagged phrases that do not appear verbatim in the input
        are dropped and the result is downgraded to warn.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequest'
            examples:
              singlePlatform:
                value:
                  content: "Lose weight fast with this guaranteed supplement!"
                  platforms: ["Facebook"]
              multiPlatform:
                value:
                  content: "Sign up now — limited time offer!"
                  platforms: ["TikTok", "Instagram", "X"]
      responses:
        '200':
          description: Scan results
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlatformResult'
        '400':
          description: Invalid input — content below 10 chars or no platforms selected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanErrorResponse'
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanErrorResponse'
        '503':
          description: Scanner not configured, upstream rate limit, or auth error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanErrorResponse'
        '504':
          description: Upstream model timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanErrorResponse'
components:
  schemas:
    GenerateRequest:
      type: object
      required: [offerText]
      additionalProperties: false
      properties:
        offerText:
          type: string
          minLength: 20
          description: Raw affiliate-offer text (network dashboard paste, email, offer sheet, etc.)
        offerUrl:
          type: string
          format: uri
          description: Optional URL to the offer's landing page; fetched server-side for verified context.
    ScanRequest:
      type: object
      required: [content, platforms]
      additionalProperties: false
      properties:
        content:
          type: string
          minLength: 10
          description: The promotional copy to scan.
        platforms:
          type: array
          minItems: 1
          items:
            type: string
            enum: [TikTok, Instagram, Facebook, Reddit, X, Pinterest, Snapchat, Discord, Telegram]
            description: Platform names. Scanner UI sends these exact spellings.
    PromoAngle:
      type: object
      required: [angle, hook, body]
      additionalProperties: false
      properties:
        angle: { type: string }
        hook: { type: string }
        body: { type: string }
    SocialPost:
      type: object
      required: [platform, character_count, text]
      additionalProperties: false
      properties:
        platform: { type: string }
        character_count: { type: integer }
        text: { type: string }
    Headline:
      type: object
      required: [variant, text]
      additionalProperties: false
      properties:
        variant: { type: string }
        text: { type: string }
    CTA:
      type: object
      required: [id, text, tone]
      additionalProperties: false
      properties:
        id: { type: string }
        text: { type: string }
        tone: { type: string }
    ComplianceNote:
      type: object
      required: [platform, note]
      additionalProperties: false
      properties:
        platform: { type: string }
        note: { type: string }
    GeneratedToolkit:
      type: object
      required: [promo_angles, social_posts, headlines, body_copy, cta_variations, compliance_notes]
      additionalProperties: false
      properties:
        promo_angles:
          type: array
          items: { $ref: '#/components/schemas/PromoAngle' }
        social_posts:
          type: array
          items: { $ref: '#/components/schemas/SocialPost' }
          description: Contains exactly 9 entries, one per platform.
        headlines:
          type: array
          items: { $ref: '#/components/schemas/Headline' }
        body_copy: { type: string }
        cta_variations:
          type: array
          items: { $ref: '#/components/schemas/CTA' }
        compliance_notes:
          type: array
          items: { $ref: '#/components/schemas/ComplianceNote' }
    PlatformResult:
      type: object
      required: [platform, status, flagged_phrases, reason, safer_rewrite]
      additionalProperties: false
      properties:
        platform: { type: string }
        status:
          type: string
          enum: [pass, warn, fail]
        flagged_phrases:
          type: array
          items: { type: string }
          description: Exact substrings of the input content. Empty when status is pass.
        reason: { type: string }
        safer_rewrite:
          type: string
          description: Empty string when status is pass.
    ErrorResponse:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error: { type: string }
    ScanErrorResponse:
      type: object
      required: [error, diagnostic]
      additionalProperties: false
      properties:
        error: { type: string }
        diagnostic: { type: string }